// src/types/users.ts
export interface InternalUser {
  id: number;
  name: string;
  username: string;
  email: string;
  image?: string;
}

export interface InternalTodo {
  id: number;
  userId: number;
  title: string;
  completed: boolean;
}

export interface UserJSONPlaceholder {
  id: number;
  name: string;
  username: string;
  email: string;
  address: {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: {
      lat: string;
      lng: string;
    };
  };
  phone: string;
  website: string;
  company: {
    name: string;
    catchPhrase: string;
    bs: string;
  };
}

export interface TodoJSONPlaceholder {
  userId: number;
  id: number;
  title: string;
  completed: boolean;
}

export interface UserDummyJSON {
  id: number;
  firstName: string;
  lastName: string;
  maidenName: string;
  age: number;
  gender: string;
  email: string;
  phone: string;
  username: string;
  password: string;
  birthDate: string;
  image: string;
  bloodGroup: string;
  height: number;
  weight: number;
  eyeColor: string;
  hair: {
    color: string;
    type: string;
  };
  ip: string;
  address: {
    address: string;
    city: string;
    state: string;
    stateCode: string;
    postalCode: string;
    coordinates: {
      lat: number;
      lng: number;
    };
    country: string;
  };
  macAddress: string;
  university: string;
  bank: {
    cardExpire: string;
    cardNumber: string;
    cardType: string;
    currency: string;
    iban: string;
  };
  company: {
    department: string;
    name: string;
    title: string;
    address: {
      address: string;
      city: string;
      state: string;
      stateCode: string;
      postalCode: string;
      coordinates: {
        lat: number;
        lng: number;
      };
      country: string;
    };
  };
  ein: string;
  ssn: string;
  userAgent: string;
  crypto: {
    coin: string;
    wallet: string;
    network: string;
  };
  role: string;
}

export type UsersDummyJSONResponse = {
  users: UserDummyJSON[];
  total: number;
  skip: number;
  limit: number;
};

export interface TodoDummyJSON {
  userId: number;
  id: number;
  todo: string;
  completed: boolean;
}

export type TodosDummyJsonResponse = {
  todos: TodoDummyJSON[];
  total: number;
  skip: number;
  limit: number;
};

export type AsyncState<T> =
  | { status: 'idle' }
  | { status: 'loading' }
  | { status: 'success'; data: T }
  | { status: 'error'; error: string };